Safari-611.1.10/LayoutTests/http/tests/navigation/target-blank-opener.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<p>Test that a new window opened via &lt;a target='_blank'&gt; with rel="noreferrer" has a null opener.</p>
<a id="testLink" href="resources/target-blank-opener-window.php" target="_blank" rel="noreferrer">Click me to test.</a>
<script>
if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.setCanOpenWindows(true);
}

document.getElementById("testLink").click();

</script>
</body>
</html>
