Safari-611.1.10/LayoutTests/http/tests/navigation/target-blank-opener-expected.txt

CONSOLE MESSAGE: Referrer:
CONSOLE MESSAGE: window.opener: PASS
Test that a new window opened via <a target='_blank'> with rel="noreferrer" has a null opener.

Click me to test.
